+++ app/layout.jsx
@@ -1,0 +1,52 @@
+import { Geist, Geist_Mono } from "next/font/google";
+import "./globals.css";
+import Link from "next/link";
+
+const geistSans = Geist({
+  variable: "--font-geist-sans",
+  subsets: ["latin"],
+});
+
+const geistMono = Geist_Mono({
+  variable: "--font-geist-mono",
+  subsets: ["latin"],
+});
+
+export const metadata = {
+  title: "Create Next App",
+  description: "Generated by create next app",
+};
+
+export default function RootLayout({ children }) {
+  return (
+    <html lang="en" version="0.0.1">
+      <body
+        className={`${geistSans.variable} ${geistMono.variable} antialiased font-[family-name:var(--font-geist-mono)]`}
+      >
+        <header className="px-4 sm:px-8 py-8">
+          <div className="pb-8">
+            <div className="flex items-center sm:items-end pb-4">
+              <div className="text-xl sm:text-5xl font-bold"><Link href={"/"}>BOOK GRAVEYARD</Link></div>
+              <div className="font-bold ml-auto text-sm sm:text-base">EST. April 24, 2025</div>
+            </div>
+            <div className="italic text-sm sm:text-base text-center sm:text-start">"Where ideas come to die. We are not trying to change the world, just trying to change your mind."</div>
+          </div>
+          <div className="block sm:flex justify-center gap-8 text-center">
+            <Link className="block" href={"/current-book-we-are-reading"}>Current Book We Are Reading</Link>
+            <span className="hidden sm:block">|</span>
+            <Link className="block" href={"/books-we-read"}>Books We Read</Link>
+            <span className="hidden sm:block">|</span>
+            <Link className="block" href={"/about-ella"}>About Ella</Link>
+            <span className="hidden sm:block">|</span>
+            <Link className="block" href={"/about-jamie"}>About Jamie</Link>
+            <span className="hidden sm:block">|</span>
+            <Link className="block" href={"/newsletter"}>Newsletter</Link>
+          </div>
+        </header>
+        <main className="px-8">
+          {children}
+        </main>
+      </body>
+    </html>
+  );
+}
