--- client/app/layout.tsx
@@ -1,34 +1,0 @@
-import { Geist, Geist_Mono } from "next/font/google";
-import { Suspense } from "react"
-import "./globals.css";
-import Header from "@/components/header";
-
-const geistSans = Geist({
-  variable: "--font-geist-sans",
-  subsets: ["latin"],
-});
-
-const geistMono = Geist_Mono({
-  variable: "--font-geist-mono",
-  subsets: ["latin"],
-});
-
-export const metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
-};
-
-export default function RootLayout({ children, }: Readonly<{ children: React.ReactNode }>) {
-  return (
-    <html lang="en" app-version="0.0.3" suppressHydrationWarning>
-      <body
-        className={`${geistSans.variable} ${geistMono.variable} antialiased font-[family-name:var(--font-geist-mono)]`}
-      >
-        <main>
-          <Header />
-          <Suspense fallback={null}>{children}</Suspense>
-        </main>
-      </body>
-    </html>
-  );
-}
