--- client/app/page.tsx
@@ -1,22 +1,0 @@
-export default function Home() {
-  return (
-    <>
-      <main>
-        <section>
-          <div className="flex justify-center pb-8">
-            <div className="max-w-lg text-center">
-              <div className="font-serif font-black text-4xl leading-tight tracking-tight text-primary">Oath</div>
-              <div className="text-justify">It is a long established fact that a reader will be distracted by the readable content of a page when looking at its layout. The point of using Lorem Ipsum is that it has a more-or-less normal distribution of letters, as opposed to using &apos;Content here, content here; making it look like readable English. Many desktop publishing packages and web pages editors now use Lorem Ipsum as their default model.</div>
-            </div>
-          </div>
-          <div className="flex justify-center">
-            <div className="w-2xl flex">
-              <div className="mr-auto">Ella Mabini</div>
-              <div className="ml-auto">Jamie Rebugio</div>
-            </div>
-          </div>
-        </section>
-      </main>
-    </>
-  );
-}
