+++ next.config.mjs
@@ -1,0 +1,6 @@
+/** @type {import('next').NextConfig} */
+const nextConfig = {
+      devIndicators: false
+};
+
+export default nextConfig;
