+++ package.json
@@ -1,24 +1,23 @@
 {
   "name": "book-graveyard",
-  "version": "0.0.3",
-  "description": "",
-  "main": "index.js",
+  "version": "0.1.0",
+  "private": true,
   "scripts": {
-    "dev:client": "npm run dev --prefix client",
-    "build:client": "npm run build --prefix client",
-    "dev:studio": "npm run dev --prefix studio",
-    "deploy:studio": "npm run deploy --prefix studio"
+    "dev": "next dev --turbopack",
+    "build": "next build",
+    "start": "next start",
+    "lint": "next lint"
   },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/romcstd/book-graveyard.git"
+  "dependencies": {
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
+    "next": "15.3.2"
   },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "type": "commonjs",
-  "bugs": {
-    "url": "https://github.com/romcstd/book-graveyard/issues"
-  },
-  "homepage": "https://github.com/romcstd/book-graveyard#readme"
+  "devDependencies": {
+    "@tailwindcss/postcss": "^4",
+    "tailwindcss": "^4",
+    "eslint": "^9",
+    "eslint-config-next": "15.3.2",
+    "@eslint/eslintrc": "^3"
+  }
 }
