--- sanity-studio/eslint.config.mjs
@@ -1,3 +1,0 @@
-import studio from '@sanity/eslint-config-studio'
-
-export default [...studio]
