--- sanity-studio/sanity.cli.ts
@@ -1,13 +1,0 @@
-import {defineCliConfig} from 'sanity/cli'
-
-export default defineCliConfig({
-  api: {
-    projectId: 'bx8fjf66',
-    dataset: 'production'
-  },
-  /**
-   * Enable auto-updates for studios.
-   * Learn more at https://www.sanity.io/docs/cli#auto-updates
-   */
-  autoUpdates: true,
-})
