--- sanity-studio/sanity.config.ts
@@ -1,18 +1,0 @@
-import {defineConfig} from 'sanity'
-import {structureTool} from 'sanity/structure'
-import {visionTool} from '@sanity/vision'
-import {schemaTypes} from './schemaTypes'
-
-export default defineConfig({
-  name: 'default',
-  title: 'Book Graveyard',
-
-  projectId: 'bx8fjf66',
-  dataset: 'production',
-
-  plugins: [structureTool(), visionTool()],
-
-  schema: {
-    types: schemaTypes,
-  },
-})
