--- sanity-studio/static/.gitkeep
@@ -1,1 +1,0 @@
-Files placed here will be served by the Sanity server under the `/static`-prefix
